Server 수정
Viewlist 추가, 패킷 관련 수정 및 추가, 클라이언트와 연동 중..

+++ Server/Server/DataBase.h
@@ -34,8 +34,8 @@ public:
 	bool PopFromStateQueue(DataBase_Event& event);
 
 	bool	GetIsLogin(std::string sID);
-	bool	GetIsLogin(unsigned short usID);
-	void	SetIsLogin(int iPlayerNum, unsigned short usID, std::string ID);
-	int		FindIsLogin(unsigned short usID, wchar_t* user_id, bool bDelete = false);
+	bool	GetIsLogin(unsigned int usID);
+	void	SetIsLogin(int iPlayerNum, unsigned int usID, std::string ID);
+	int		FindIsLogin(unsigned int usID, wchar_t* user_id, bool bDelete = false);
 };
